#26:Using host network for execution of exasclt
In case of exasclt upload the user might try to push image to a custom network path. For that we need to execute the exasclt docker instance in the host network, not the isolated docker network.

Also, execute tests within container, by using exaslct script instead of starter_scripts/exaslct_without_poetry.sh