#164: Fixed default values for click parameters of type multiple=true
RCA: There is an inconsistency in click, when using parameter of type multiple=true. The runtime object of the click command, indicates a default-value=None; however, when invoking the command without the respective parameter, click passes an empty tuple.

Solution: Implement a special handling when the default value of a parameter is None, but the parameter is of type multiple=true